COVER: Update the Netduino 2 Machine
---CHANGELOG---
PUBLIC
This patchset continues with the Netduino 2 and STM32F205 SoC
work.

This patch series makes a small change to the STM32F2xx
SoC to tidy up the code.

Next a feature is added to the STM32F2xx timer to display the
PWM duty cycle, when debugging is enabled.

Then the STM32F2xx SPI and ADC devices are added and connected
to the STM32F205 SoC.

Finally the maintainers file is updated to add myself as the
maintainer for the Netdunio 2 and STM32F2xx.

V8:
 - Use the qdev_init_gpio_in() function to create in irqs for the OR device.
V6:
 - Add vmstate to the new OR IRQ device
V5:
 - Add a irq ORing function
V4:
 - Add VMState to the new devices
 - Remove rand() function
V3:
 - Rebase
V2:
 - Update based on Peter C's coments
 - Rebase
 - Create an ADC folder for the ADC device
---END---

Signed-off-by: Alistair Francis <alistair@alistair23.me>